# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Slab Code Projects APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Signup user",
                "description": "Signup user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User that should be signup",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "nombre",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "email",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "password",
                                    "type": "string"
                                },
                                "roles": {
                                    "description": "roles",
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Registrar usuario",
                                    "type": "string",
                                    "example": "Successfully created user!"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Registrar usuario",
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "description": "errors object",
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Login user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User that should be login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "email",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "password",
                                    "type": "string"
                                },
                                "remember_me": {
                                    "description": "remember password",
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "access_token": {
                                    "description": "Inicio de sesi\u00f3n y generaci\u00f3n de token",
                                    "type": "string",
                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9....."
                                },
                                "token_type": {
                                    "description": "Inicio de sesi\u00f3n y generaci\u00f3n de token",
                                    "type": "string",
                                    "example": "Bearer"
                                },
                                "expires_at": {
                                    "description": "Inicio de sesi\u00f3n y generaci\u00f3n de token",
                                    "type": "string",
                                    "example": "2020-12-18 02:58:28"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/user/{id}/disable": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Disable user",
                "description": "Disable user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of User",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Disable User",
                                    "type": "string",
                                    "example": "User disable successfully"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
}